refactor: extract shared seeder logic to SeederHelper (fix Sonar duplication)
DefaultSystemUserSeeder and IntegrationBootstrapSeeder had a 63-line
duplicated block (EnsureUserWithAllPermissions logic). SonarCloud
reported new_duplicated_lines_density = 6.1% (threshold: 3%), failing
the Quality Gate.

Extract the common upsert-user-with-all-permissions flow into a new
internal SeederHelper class. Both seeders now delegate to it, preserving
their public API (Email, ResolveCredential, Ensure*Async).

Made-with: Cursor
